fixup! refactor(json_view): simplify function names
